+++ src/App.tsx
@@ -4,8 +4,8 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { Login } from './pages/login';
 import { OfferPage } from './pages/offer';
 import { Favorites } from './pages/favorites';
-import { PrivateRoute } from './components/privateRoute';
-import { PlacesCount } from './const/placesCount.ts';
+import { PrivateRoute } from './components/private-route';
+import { PlacesCount } from './const/places-count.ts';
 import { Error } from './components/error';
 import { OfferType } from './types';
 import { MockCity, Points } from './mocks';
