+++ src/components/card/index.ts
@@ -1,1 +1,1 @@
-export * from './Card.tsx';
+export * from './card.tsx';
