+++ src/components/error/index.ts
@@ -1,1 +1,1 @@
-export * from './Error.tsx';
+export * from './error.tsx';
