+++ src/components/favorite-empty/index.ts
@@ -1,0 +1,1 @@
+export * from './favorite-empty.tsx';
