--- src/components/favoriteEmpty/index.ts
@@ -1,1 +1,0 @@
-export * from './FavoriteEmpty.tsx';
