+++ src/components/footer/index.ts
@@ -1,1 +1,1 @@
-export * from './Footer.tsx';
+export * from './footer.tsx';
