+++ src/components/header/index.ts
@@ -1,1 +1,1 @@
-export * from './Header.tsx';
+export * from './header.tsx';
