+++ src/components/map/index.ts
@@ -1,1 +1,1 @@
-export * from './Map.tsx';
+export * from './map.tsx';
