+++ src/components/map/map.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useRef } from 'react';
 import { Location, OfferType } from '../../types/';
 import { Icon, layerGroup, Marker } from 'leaflet';
-import { URL_MARKER_ACTIVE, URL_MARKER_DEFAULT } from '../../const/urlMarkers.ts';
-import useMap from '../../hooks/useMap.ts';
+import { URL_MARKER_ACTIVE, URL_MARKER_DEFAULT } from '../../const/url-markers.ts';
+import useMap from '../../hooks/use-map.ts';
 
 type MapProps = {
   city: Location;
