+++ src/components/offers-list/index.ts
@@ -1,0 +1,1 @@
+export * from './offers-list.tsx';
