--- src/components/offersList/index.ts
@@ -1,1 +1,0 @@
-export * from './OffersList.tsx';
