--- src/components/privateRoute/index.ts
@@ -1,1 +1,0 @@
-export * from './PrivateRoute.tsx';
