+++ src/components/reviews-form/index.ts
@@ -1,0 +1,1 @@
+export * from './reviews-form.tsx';
