+++ src/components/reviews-list/index.ts
@@ -1,0 +1,1 @@
+export * from './reviews-list.tsx';
