--- src/components/reviewsForm/index.ts
@@ -1,1 +1,0 @@
-export * from './ReviewsForm.tsx';
