--- src/components/reviewsList/index.ts
@@ -1,1 +1,0 @@
-export * from './ReviewsList.tsx';
