+++ src/mocks/index.ts
@@ -1,6 +1,6 @@
-export * from './mockOffers.ts';
-export * from './mockCity.ts';
+export * from './mock-offers.ts';
+export * from './mock-city.ts';
 export * from './points.ts';
-export * from './mockReviews.ts';
-export * from './mockNearby.ts';
+export * from './mock-reviews.ts';
+export * from './mock-nearby.ts';
