+++ src/mocks/mock-offers.ts
