+++ src/pages/favorites/index.ts
@@ -1,1 +1,1 @@
-export * from './Favorites.tsx';
+export * from './favorites.tsx';
