+++ src/pages/login/index.ts
@@ -1,1 +1,1 @@
-export * from './Login.tsx';
+export * from './login.tsx';
