+++ src/pages/main/index.ts
@@ -1,1 +1,1 @@
-export * from './Main.tsx';
+export * from './main.tsx';
