+++ src/pages/offer-not-logged/index.ts
@@ -1,0 +1,1 @@
+export * from './offer-not-logged.tsx';
