--- src/pages/offerNotLogged/index.ts
@@ -1,1 +1,0 @@
-export * from './OfferNotLogged.tsx';
